Rename activitysources.js, resultsmodifiers.js
